tighten up flow types, add build:flow to add flow files to lib.
