+++ package.json
@@ -17,9 +17,10 @@
     "lint:fix": "npm run lint -- --fix",
     "build:css": "node-sass css/styles.scss dist/style.css",
     "build:lib": "cross-env BABEL_ENV=commonjs babel src --out-dir lib",
+    "build:flow": "flow-copy-source -v -i '**/__tests__/**' src lib",
     "build:umd": "cross-env BABEL_ENV=commonjs NODE_ENV=development webpack src/index.js dist/redux-form-gen.js",
     "build:umd:min": "cross-env BABEL_ENV=commonjs NODE_ENV=production webpack src/index.js dist/redux-form-gen.min.js",
-    "build": "npm run build:lib && npm run build:umd && npm run build:umd:min && npm run build:css",
+    "build": "npm run build:lib && npm run build:flow && npm run build:umd && npm run build:umd:min && npm run build:css",
     "deploy-storybook": "storybook-to-ghpages",
     "test": "cross-env BABEL_ENV=commonjs jest --config .jestconfig.json --verbose",
     "test:cov": "npm test -- --coverage",
@@ -80,6 +81,7 @@
     "eslint-plugin-react": "^7.4.0",
     "eslint-plugin-standard": "^3.0.1",
     "flow-bin": "^0.59.0",
+    "flow-copy-source": "^1.2.1",
     "husky": "^0.14.3",
     "jest": "^21.0.2",
     "node-sass": "^4.5.0",
