+++ src/FormGenerator.js
@@ -30,7 +30,6 @@ class FormGenerator extends Component<Props, State> {
     customQuestionProps: PropTypes.object, // object key = questionId, value = an object of props for the specified questionId's
     visibleDepth: PropTypes.string, // dot-notation string, used to rendering fields after this path (startsWith)
     customFieldTypes: PropTypes.objectOf(PropTypes.func),
-    sectionArrayIndex: PropTypes.number,
     display: PropTypes.oneOf(['stacked', 'inline']),
     disabled: PropTypes.bool // if true, disables all fields in the form
   };
