+++ src/FormGenerator.types.js
@@ -3,14 +3,14 @@ import type {LookupTable} from './utils.types';
 
 export type Props = {
   fields: FieldsType,
-  customQuestionProps: {
+  customQuestionProps?: {
     [key: string]: Object
   },
   visibleDepth?: string,
   customFieldTypes?: CustomFieldTypes,
   display?: 'stacked' | 'inline',
   disabled?: boolean,
-  lookupTable?: LookupTable
+  lookupTable: LookupTable
 };
 
 export type State = {
